Fix energy bar was not correctly rendered if more than 40 M FE/t was stored

+++ src/main/java/me/jddev0/ep/screen/base/IEnergyStorageMenu.java
@@ -8,7 +8,7 @@ public interface IEnergyStorageMenu extends IUpgradeModuleMenu {
         int energy = getEnergy();
         int capacity = getCapacity();
 
-        return Math.min(energyMeterHeight, (energy == 0 || capacity == 0)?0:Math.max(1, energy * energyMeterHeight / capacity));
+        return Math.min(energyMeterHeight, (energy == 0 || capacity == 0)?0:(int)Math.max(1, (long)energy * energyMeterHeight / capacity));
     }
 
     default int getEnergyIndicatorBarValue() {
@@ -27,6 +27,6 @@ public interface IEnergyStorageMenu extends IUpgradeModuleMenu {
         int energyPerTick = getEnergyPerTickBarValue();
         int capacity = getCapacity();
 
-        return (energyPerTick <= 0 || capacity == 0)?0:(Math.min(energyPerTick, capacity - 1) *energyMeterHeight / capacity + 1);
+        return (energyPerTick <= 0 || capacity == 0)?0:(int)((long)Math.min(energyPerTick, capacity - 1) * energyMeterHeight / capacity + 1);
     }
 }
